% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{and_helper}
\alias{and_helper}
\title{and standardization helper function}
\usage{
and_helper(x)
}
\arguments{
\item{x}{A character vector of titles.}
}
\description{
operates on a singular title level
splits it, checks that if both/all sub titles are valid titles,
then returns a boolean whether and was used as a title separator
later lapply-ed for standardize_and usage

as opposed to & its default is false
}
