% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-classifier-functions.R
\name{apply_cleaning}
\alias{apply_cleaning}
\title{apply cleaning function}
\usage{
apply_cleaning(title.text)
}
\description{
\code{apply_cleaning} applies cleaning steps (capitalization, punctuation and number
removal, leading spaces, etc.) on a title text. For example, it would convert
an input of "  pResid." into "PRESID". For future work, the goal is to extract
dates and role statuses during this cleaning step in order to improve
efficiency, but currently that is a separate step. The function takes in an
input of a string and outputs a string.
}
