% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-split-titles.R
\name{apply_misc_split_rules}
\alias{apply_misc_split_rules}
\title{misc split rules for edge cases}
\usage{
apply_misc_split_rules(x)
}
\arguments{
\item{x}{A character vector of titles.}
}
\description{
Add some additional split rules for cases
not identified by previous steps.
}
