% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-classifier-functions.R
\name{apply_substitutes}
\alias{apply_substitutes}
\title{apply substitutes function}
\usage{
apply_substitutes(title.text)
}
\description{
wrapper function

apply custom dictionary of common abbreviations and misspellings
Co (co-chair) vs Co. (coordinator)
unmingle()  # use spell check to separate 2 words missing a space:  executivedirector
spellcheck()   # only apply to words over certain length, and only make substitutions of 1 or 2 small changes (extra letter, flipped order of letters, missing letter, wrong letter)

this is probably to take the place of apply_cleaning
Note: not yet written yet
}
