% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-title-classifier-functions.R
\name{build_split_titles}
\alias{build_split_titles}
\title{build split titles function}
\usage{
build_split_titles(comp.data)
}
\description{
basically doing the smae thing as build standard titles but on a split title
wrapper function
it will be applied on a data frame

simply splitting on "&"
assuming it's already cleaned well

runtime on 100,000 titles ~ 1 hour
}
