% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-classifier-functions.R
\name{build_standard_titles}
\alias{build_standard_titles}
\title{build standard titles function
DEPRECATED}
\usage{
build_standard_titles(comp.table)
}
\description{
\code{build_standard_titles} takes in a compensation table and cleans up the titles.
It does some basic data extraction from the titles (including date extraction,
standardization, and more), but there is still more work to do.
Note: this function was specifically designed for jesse's title-test-run.csv,
but it has been slightly modified so that it works more generally from part
vii table 01 builds.
This test function combines all the functions from this file to manipulate
a data frame. It takes in a data frame and outputs a filtered data frame.
}
