% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-categorize-titles.R
\name{categorize_ceo}
\alias{categorize_ceo}
\title{categorize ceo function}
\usage{
categorize_ceo(df)
}
\arguments{
\item{df}{A compensation data frame.}
}
\value{
Returns a data frame with a CEO boolean (1=CEO,0=other).
}
\description{
categorizing ceo's and adding a flag
}
