% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize-titles-new.R
\name{categorize_company_leader}
\alias{categorize_company_leader}
\title{categorize company leader function}
\usage{
categorize_company_leader(comp.data)
}
\description{
utilizes ceo and adds in president, chair, and managing director for those
without explicit/clear ceo positions
}
