% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-gen-status-codes.R
\name{categorize_qualifiers}
\alias{categorize_qualifiers}
\title{categorize qualifiers function}
\usage{
categorize_qualifiers(comp.data, title = "TitleTxt6")
}
\description{
takes in a data frame and categorizes the role statuses
future, former, interim, current (current is default, no status for that)
}
