% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize-titles-new.R
\name{categorize_titles_new}
\alias{categorize_titles_new}
\title{categorize titles function
will be updated version
will be a wrapper function}
\usage{
categorize_titles_new(comp.data)
}
\description{
categorizes the titles in comp.data
will work on titletxt4
}
