% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunction-cleaning.R
\name{clean_conjunctions}
\alias{clean_conjunctions}
\title{clean conjunctions function}
\usage{
clean_conjunctions(title.text)
}
\description{
wrapper for cleaning conjunctions (all the standardizations come in here)
}
