% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-title-classifier-functions.R
\name{clean_dates}
\alias{clean_dates}
\title{clean dates wrapper function, takes in a data frame}
\usage{
clean_dates(comp.data)
}
\description{
cleans the dates by first identifying date, setting a flag if it finds one,
then removing the date if it is present
}
