% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-title-classifier-functions.R
\name{clean_empty_cells}
\alias{clean_empty_cells}
\title{clean empty cells function}
\usage{
clean_empty_cells(comp.data)
}
\description{
removes any title rows without any meaningful information in the title cell,
i.e. titles cells without any titles in them
#we skip over titles that may have had valuable info removed like emeritus or
at large
}
