% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{comma_helper}
\alias{comma_helper}
\title{standardize comma helper function}
\usage{
comma_helper(x)
}
\arguments{
\item{x}{A character vector of titles.}
}
\description{
returns 0,1,2 depending on whether a comma is used as a title separator,
word separator, or just extraneous
operates on the atomic vector level
}
