% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-classifier-functions.R
\name{comma_split}
\alias{comma_split}
\title{comma split function
DEPRECATED}
\usage{
comma_split(title.text)
}
\description{
\code{comma_split} checks if using a comma as a title delineator works (if appl.).
It works by creating a vector of the comma split titles, and checking that
the standardization of the second title is valid. It does this to ensure
cases like "VP, Sales and Marketing" aren't broken up into "VP" and "Sales
and Marketing", but cases like "CFO, Treasurer" are split into "CFO" and
"Treasurer". The standardized list of titles is pretty general, and could be
changed/improved as more edge cases are discovered.
}
