% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{condense_abbreviations}
\alias{condense_abbreviations}
\title{condense c-suite abbreviations function}
\usage{
condense_abbreviations(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses spaced out c-suite abbreviations
e.g. C E O --> CEO

also includes miscellaneous executive director and key employee substitutions
}
