% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remove-dates.R
\name{convert_ordinal}
\alias{convert_ordinal}
\title{convert ordinal numbers function}
\usage{
convert_ordinal(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
converts ordinal numbers (1st) to their alphabetic counterpart (first)
}
