% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-classifier-functions.R
\name{extract_date}
\alias{extract_date}
\title{extract date function}
\usage{
extract_date(title.text)
}
\description{
\code{extract_date} extracts the date information from an unfiltered title string,
if present. It works by first checking parentheticals for numbers, then
"XX/XX/XXXX" dates, then spelled out months. It extracts the nearby info using
a flexible RegEx matching scheme. The function takes in a title text string
and outputs a length 2 vector with the first element being the date string,
and the second being the number of detected dates in the string. If there is
no date present, the return vector has the first element NA and the second 0.
}
