% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-codes.R
\name{filter_sched_o}
\alias{filter_sched_o}
\title{filter schedule o function}
\usage{
filter_sched_o(TitleTxt)
}
\description{
add binary code for “see schedule o” comments and remove such text
also includes see attached and abbreviations of sched o
}
