% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_academics}
\alias{fix_academics}
\title{standardize versions of 'academics'}
\usage{
fix_academics(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses academics abbreviations to a standardized form
}
