% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_admin}
\alias{fix_admin}
\title{standardize versions of 'admin'}
\usage{
fix_admin(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses administration/administrator/administrating
abbreviations to a standardized form
}
