% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_admissions}
\alias{fix_admissions}
\title{standardize versions of 'admissions'}
\usage{
fix_admissions(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses admissions abbreviations to a standardized form
}
