% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_advancement}
\alias{fix_advancement}
\title{standardize versions of 'advancement'}
\usage{
fix_advancement(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses advancement abbreviations to a standardized form
}
