% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_assistant}
\alias{fix_assistant}
\title{standardize versions of 'assistant'}
\usage{
fix_assistant(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses assistant abbreviations to a standardized form
also condenses associate
}
