% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_atlarge}
\alias{fix_atlarge}
\title{standardize versions of 'at large'}
\usage{
fix_atlarge(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses at large abbreviations to a standardized form
}
