% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_board}
\alias{fix_board}
\title{standardize versions of 'board'}
\usage{
fix_board(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses board abbreviations to a standardized form
}
