% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_business}
\alias{fix_business}
\title{standardize versions of 'business'}
\usage{
fix_business(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses business abbreviations to a standardized form
}
