% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_deputy}
\alias{fix_deputy}
\title{standardize versions of 'deputy'}
\usage{
fix_deputy(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses deputy abbreviations to a standardized form
}
