% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_director}
\alias{fix_director}
\title{standardize versions of 'director'}
\usage{
fix_director(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses director abbreviations to a standardized form
}
