% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_emeritus}
\alias{fix_emeritus}
\title{standardize versions of 'emeritus'}
\usage{
fix_emeritus(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses emeritus abbreviations to a standardized form
}
