% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_executive}
\alias{fix_executive}
\title{standardize versions of 'executive'}
\usage{
fix_executive(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses executive abbreviations to a standardized form
}
