% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_finance}
\alias{fix_finance}
\title{standardize versions of 'finance'}
\usage{
fix_finance(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses finance abbreviations to a standardized form
financial is included
}
