% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_general}
\alias{fix_general}
\title{standardize versions of 'general'}
\usage{
fix_general(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses general abbreviations to a standardized form
}
