% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_hr}
\alias{fix_hr}
\title{standardize versions of 'human resources'}
\usage{
fix_hr(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses hr abbreviations to a standardized form
}
