% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_information}
\alias{fix_information}
\title{standardize versions of 'information'}
\usage{
fix_information(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses info abbreviations to a standardized form
}
