% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_intelligence}
\alias{fix_intelligence}
\title{standardize versions of 'intelligence'}
\usage{
fix_intelligence(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses intel abbreviations to a standardized form
}
