% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_marketing}
\alias{fix_marketing}
\title{standardize versions of 'marketing'}
\usage{
fix_marketing(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses marketing abbreviations to a standardized form
}
