% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_member}
\alias{fix_member}
\title{standardize versions of 'member'}
\usage{
fix_member(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses member abbreviations to a standardized form
}
