% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_of}
\alias{fix_of}
\title{fix "of" function}
\usage{
fix_of(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
inserts an "of" in between the title and subject if not present and needed
e.g. would insert "of" in "VP Operations" --> "VP of Operations"

currently occurs after title standardization
(but can potentially occur before too)
}
