% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_officer}
\alias{fix_officer}
\title{standardize versions of 'officer'}
\usage{
fix_officer(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses officer abbreviations to a standardized form
}
