% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_operations}
\alias{fix_operations}
\title{standardize versions of 'operations'}
\usage{
fix_operations(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses operations abbreviations to a standardized form
function leaves "operating" alone --> does not convert to operations
}
