% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_planning}
\alias{fix_planning}
\title{standardize versions of 'planning'}
\usage{
fix_planning(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses planning abbreviations to a standardized form
includes planned
}
