% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_programs}
\alias{fix_programs}
\title{standardize versions of 'programs'}
\usage{
fix_programs(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses programs abbreviations to a standardized form
skips over programming
}
