% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_projects}
\alias{fix_projects}
\title{standardize versions of 'projects'}
\usage{
fix_projects(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses projects abbreviations to a standardized form
}
