% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_public}
\alias{fix_public}
\title{standardize versions of 'public'}
\usage{
fix_public(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses public abbreviations to a standardized form
}
