% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_relations}
\alias{fix_relations}
\title{standardize versions of 'relations'}
\usage{
fix_relations(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses relations abbreviations to a standardized form
}
