% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_secretary}
\alias{fix_secretary}
\title{standardize versions of 'secretary'}
\usage{
fix_secretary(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses secretary abbreviations to a standardized form
}
