% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_senior}
\alias{fix_senior}
\title{standardize versions of 'seniority'}
\usage{
fix_senior(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses senior/junior abbreviations to a standardized form
}
