% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_strategy}
\alias{fix_strategy}
\title{standardize versions of 'strategy'}
\usage{
fix_strategy(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses strategy abbreviations to a standardized form
also condenses strategic to strategy
}
