% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_systems}
\alias{fix_systems}
\title{standardize versions of 'systems'}
\usage{
fix_systems(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses systems abbreviations to a standardized form
}
