% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_technology}
\alias{fix_technology}
\title{standardize versions of 'technology'}
\usage{
fix_technology(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses tech abbreviations to a standardized form
}
