% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_treasurer}
\alias{fix_treasurer}
\title{standardize versions of treasurer helper function}
\usage{
fix_treasurer(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses treasurer abbreviations to a standardized form
}
