% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_vice}
\alias{fix_vice}
\title{standardize versions of 'vice'}
\usage{
fix_vice(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
condenses vice president abbreviations to a standardized form
}
