% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-gen-status-codes.R
\name{flag_and_remove}
\alias{flag_and_remove}
\title{create a status code flag and remove string from title}
\usage{
flag_and_remove(
  df,
  title = "TitleTxt6",
  s.code = "FORMER",
  gs_status_codes = NULL
)
}
\arguments{
\item{df}{A compensation dataframe}

\item{title}{Which version of the title string to use (defaults to "TitleTxt6")}

\item{s.code}{Any of the unique status.qualifier strings from df.status (defaults to "FORMER")}

\item{gs_status_codes}{Optional status-code crosswalk data frame; if \code{NULL}, loaded from the package data.}
}
\description{
Search for variants of a status code (in the status.codes crosswalk),
create a boolean flag in the dataset, and remove the status qualifier from the title.
The flag variable is named SCODE.X (FORMER.X, INTERIM.X, etc.).

In cases where the status is the full title (e.g. 'ex officio') replace the variant
with the standardized version and keep it.
}
\examples{
\dontrun{
x <-
  c( "IMMEDIATE PAST CHAIR",
     "FORMER CEO (EXIT 12/31/17)",
     "PAST PRESIDENT",
     "OUTGOING CEO",
     "PRESIDENT ELECT",
     "BOARD MEMBER (START OCT)",
     "TREASURER - BEGINNING 8/2018",
     "DIRECTOR (AS OF 2/22/18)",
     "INTERIM PRES",
     "ACTING DIRECTOR",
     "PAST INTERIM PRESIDENT" )

remove_date( x )
}

}
