% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-title-classifier-functions.R
\name{format_comp_df}
\alias{format_comp_df}
\title{Format the compensation table data frame (from 990 form, Part VII, Table 01)}
\usage{
format_comp_df(comp.data)
}
\description{
\code{format_comp_df} returns a processed data frame with renamed column titles
and cleaned up fields from the raw table extracted from PVII T01 build. Also,
this function removes duplicated lines if present. Both the input and output
are data frames. Ultimately this function provides takes in raw data and
returns a data frame with understandable information about DTK individuals.
}
