% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-tables.R
\name{gen_helpful_tables}
\alias{gen_helpful_tables}
\title{gen helpful tables function}
\usage{
gen_helpful_tables(comp.table)
}
\arguments{
\item{comp.table}{A Part VII compensation data frame.}
}
\description{
generates and saves some helpful tables from a compensation table
including raw and cleaned titles including some specific string.
these list include titles with "co", "of", "to", ",", "/", "&", and "and".
they also include titles with ordinal numbers (1st, second, 3rd, etc.) and
c-suite titles (CEO, Chief Financial Officer).

The titles are saved into the /test-tables directory and have the file
naming convention of "titles with ~specific string~ ~date~".
They are in csv format.
}
