% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-gen-taxonomy.R
\name{gen_taxonomy}
\alias{gen_taxonomy}
\title{generate taxonomy function}
\usage{
gen_taxonomy(comp.data)
}
\arguments{
\item{comp.data}{A Part VII compensation data frame.}
}
\description{
TODO not written yet, but we will generate the taxonomy using the title-taxonomy crosswalk
no longer manipulating the titles
}
