% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-tables.R
\name{get_title_dump}
\alias{get_title_dump}
\title{get title dump function}
\usage{
get_title_dump(comp.table, target.word)
}
\arguments{
\item{comp.table}{A Part VII compensation data frame.}

\item{target.word}{A word, or vector of words, to search for within titles.}
}
\description{
generate all the titles that contain a given word and output a dataframe
including the original raw title and its cleaned version
this function works for individual words as well as word lists
}
