% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remove-dates.R
\name{has_date}
\alias{has_date}
\title{identify dates}
\usage{
has_date(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
returns logical: TRUE if a string contains a date
}
