% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-standardize-df.R
\name{pre_clean}
\alias{pre_clean}
\title{pre cleaning function}
\usage{
pre_clean(x)
}
\arguments{
\item{x}{A character vector of titles.}
}
\description{
gets rid of meaningless punctuation (like periods),
converts titles to uppercase form
}
