% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-split-titles.R
\name{remove_first_split}
\alias{remove_first_split}
\title{remove first split function}
\usage{
remove_first_split(x)
}
\arguments{
\item{x}{A character vector of titles.}
}
\description{
removes the first occurrence of a split title
}
