% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{remove_trailing_conjunctions}
\alias{remove_trailing_conjunctions}
\title{remove trailing conjunctions}
\usage{
remove_trailing_conjunctions(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
removes any trailing conjunctions (and, of, to)
as well as any leading conjunctions
}
