% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-standardize-titles.R
\name{replace_board}
\alias{replace_board}
\title{replace board function}
\usage{
replace_board(TitleTxt)
}
\description{
replaces all instances of titles that could be a board position
including finance director and such
}
