% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-standardize-titles.R
\name{replace_ceo}
\alias{replace_ceo}
\title{replace ceo function}
\usage{
replace_ceo(TitleTxt, weekly.hours, total.pay)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}

\item{weekly.hours}{Numeric vector of average weekly hours.}

\item{total.pay}{Numeric vector of total compensation.}
}
\description{
replaces all instances of titles that could be CEO
}
