% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{simplify_clevels}
\alias{simplify_clevels}
\title{standardize title mapping function}
\usage{
simplify_clevels(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
maps long names to abbreviations if CEO, COO, or CFO
maps executive director to CEO
}
