% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-categorize-titles.R
\name{simplify_varnames}
\alias{simplify_varnames}
\title{rename variables in the dataset}
\usage{
simplify_varnames(df)
}
\arguments{
\item{df}{A Part VII compensation data frame.}
}
\description{
simplify variable names
}
