% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{slash_helper}
\alias{slash_helper}
\title{standardize slash helper function}
\usage{
slash_helper(x)
}
\arguments{
\item{x}{A character vector of titles.}
}
\description{
checks for a slash and replaces it
with an ampersand;
operates on single title
}
