% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{spellcheck}
\alias{spellcheck}
\title{spell check function}
\usage{
spellcheck(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
correct small spelling mistakes
right now only correcting president, treasurer, and trustee misspellings

pretty slow function right now...
}
