% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-classifier-functions.R
\name{split_compound_title}
\alias{split_compound_title}
\title{Split compound titles
DEPRECATED}
\usage{
split_compound_title(title.text)
}
\description{
\code{split_compound_title} splits compound titles into separate strings if evident.
For example, it would split "Treasurer/Secretary" into a vector of
strings containing ("Treasurer", "Secretary"). The production rules for title
delineation are "/", ";", "&", "AND", and "," in certain cases ("," is the
weakest delineator, so it is only used if there is a guarantee that the second
title is standardized). The input is a string, and the output is a vector
with variable length.
}
