% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-split-titles.R
\name{split_title}
\alias{split_title}
\title{split title function}
\usage{
split_title(TitleTxt)
}
\description{
splits compound titles using & as a separator,
returning a list of possible list of titles
}
