% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-split-titles.R
\name{split_titles}
\alias{split_titles}
\title{split titles wrapper function}
\usage{
split_titles(df, title = "TitleTxt3")
}
\arguments{
\item{df}{A Part VII compensation data frame.}

\item{title}{Name of the title column to operate on.}
}
\description{
creates a flag for a split, and then builds out a new data frame
(combining it back together) later --> should run much faster

runtime compared to old code is about several magnitudes faster than before
}
