% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{stand_titles}
\alias{stand_titles}
\title{standardize title mapping function}
\usage{
stand_titles(TitleTxt)
}
\description{
maps long names to abbreviations if CEO, COO, or CFO
maps executive director to CEO
}
