% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{standardize_and}
\alias{standardize_and}
\title{standardize "and" function}
\usage{
standardize_and(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
detects whether "and" or & is a title separator
use & as separator, "and" as regular?
assume our dates are already extracted

& is separator, and is regular
}
