% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{standardize_conj}
\alias{standardize_conj}
\title{split conjunctions function}
\usage{
standardize_conj(comp.data, title = "TitleTxt2")
}
\arguments{
\item{comp.data}{A Part VII compensation data frame.}

\item{title}{Name of the title column to operate on.}
}
\description{
wrapper for cleaning conjunctions (all the standardizations come in here)
}
