% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-codes.R
\name{standardize_current}
\alias{standardize_current}
\title{standardize current words function}
\usage{
standardize_current(title.text)
}
\description{
standardizes all synonyms for current in the title text and then removes them

works as a catch all for titles that don't get categorized using the other
3 role statuses
}
