% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-codes.R
\name{standardize_former}
\alias{standardize_former}
\title{standardize former words function}
\usage{
standardize_former(title.text)
}
\description{
standardizes all synonyms for former in the title text and then removes them
}
