% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-codes.R
\name{standardize_future}
\alias{standardize_future}
\title{standardize future words function}
\usage{
standardize_future(title.text)
}
\description{
standardizes all synonyms for future in the title text and then removes them
}
