% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-codes.R
\name{standardize_interim}
\alias{standardize_interim}
\title{standardize interim words function}
\usage{
standardize_interim(title.text)
}
\description{
standardizes all synonyms for interim in the title text and then removes them
}
