% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{standardize_separator}
\alias{standardize_separator}
\title{standardize separator function}
\usage{
standardize_separator(x)
}
\arguments{
\item{x}{A character vector of titles.}
}
\description{
standardize the separator for distinguishing titles
possible separators are & ; and \ (rarely used but sometimes)
note: / and , already caught
They all get mapped to &
}
