% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{standardize_slash}
\alias{standardize_slash}
\title{standardize slash function}
\usage{
standardize_slash(TitleTxt)
}
\arguments{
\item{TitleTxt}{A character vector of titles.}
}
\description{
replaces slash with '&'
}
