% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunction-cleaning.R
\name{standardize_space}
\alias{standardize_space}
\title{standardize space function}
\usage{
standardize_space(title.text)
}
\description{
only really used with unmingle (otherwise it's too volatile)
}
