% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{standardize_spelling}
\alias{standardize_spelling}
\title{standardize spelling wrapper function}
\usage{
standardize_spelling(comp.data, title = "TitleTxt4")
}
\arguments{
\item{comp.data}{A Part VII compensation data frame.}

\item{title}{Name of the title column to operate on.}
}
\description{
apply custom dictionary of common abbreviations and misspellings
to standardize titles
}
