% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-gen-status-codes.R
\name{standardize_status}
\alias{standardize_status}
\title{replace status variant with the standardized version}
\usage{
standardize_status(x, s.code, variants)
}
\arguments{
\item{x}{Vector of titles}

\item{s.code}{Any of the unique status.qualifier strings from df.status}

\item{variants}{Regex alternation of status-code variants (from \code{get_variants()}).}
}
\description{
Search for variants of a status code and replace the variant
with the standardized version of the status code.
}
