% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status.mapping}
\alias{status.mapping}
\title{Status-code mapping table.}
\description{
A data frame mapping status-qualifier variants (e.g. "ACTING",
"EMERITUS") to standardized status codes (e.g. "INTERIM", "FORMER") used in
the \code{gen_status_codes()} step.
}
