% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_admin}
\alias{substitute_admin}
\title{substitute admin helper function}
\usage{
substitute_admin(TitleTxt)
}
\description{
condenses administration/administrator/administrating
abbreviations to a standardized form
}
