% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_advancement}
\alias{substitute_advancement}
\title{substitute advancement helper function}
\usage{
substitute_advancement(title.text)
}
\description{
condenses advancement abbreviations to a standardized form
}
