% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_assistant}
\alias{substitute_assistant}
\title{substitute assistant helper function}
\usage{
substitute_assistant(TitleTxt)
}
\description{
condenses assistant abbreviations to a standardized form
also condenses associate
}
