% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_atlarge}
\alias{substitute_atlarge}
\title{substitute at large helper function}
\usage{
substitute_atlarge(TitleTxt)
}
\description{
condenses at large abbreviations to a standardized form
}
