% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_board}
\alias{substitute_board}
\title{substitute board helper function}
\usage{
substitute_board(TitleTxt)
}
\description{
condenses board abbreviations to a standardized form
}
