% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_business}
\alias{substitute_business}
\title{substitute business helper function}
\usage{
substitute_business(TitleTxt)
}
\description{
condenses business abbreviations to a standardized form
}
