% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_comm}
\alias{substitute_comm}
\title{substitute comm helper functions}
\usage{
substitute_comm(title.text)
}
\description{
condenses communication and committee abbreviations to a standardized form
}
