% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_coordinator}
\alias{substitute_coordinator}
\title{substitute coordinator helper function}
\usage{
substitute_coordinator(title.text)
}
\description{
condenses coordinator abbreviations to a standardized form
note: instead of standardizing "CO" at the end of a title to coordinator,
it is instead mapped to "COMMITTEE"
}
