% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_corresponding}
\alias{substitute_corresponding}
\title{substitute corresponding helper function}
\usage{
substitute_corresponding(TitleTxt)
}
\description{
condenses corresponding abbreviations to a standardized form
}
