% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_deputy}
\alias{substitute_deputy}
\title{substitute deputy helper function}
\usage{
substitute_deputy(TitleTxt)
}
\description{
condenses deputy abbreviations to a standardized form
}
