% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_development}
\alias{substitute_development}
\title{substitute development helper function}
\usage{
substitute_development(TitleTxt)
}
\description{
condenses development abbreviations to a standardized form
}
