% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_director}
\alias{substitute_director}
\title{substitute director helper functions}
\usage{
substitute_director(TitleTxt)
}
\description{
condenses director abbreviations to a standardized form
}
