% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_emeritus}
\alias{substitute_emeritus}
\title{substitute emeritus helper function}
\usage{
substitute_emeritus(title.text)
}
\description{
condenses emeritus abbreviations to a standardized form
}
