% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_enrollment}
\alias{substitute_enrollment}
\title{substitute enrollment helper function}
\usage{
substitute_enrollment(title.text)
}
\description{
condenses enrollment abbreviations to a standardized form
}
