% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_executive}
\alias{substitute_executive}
\title{substitute executive helper function}
\usage{
substitute_executive(TitleTxt)
}
\description{
condenses executive abbreviations to a standardized form
}
