% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_exofficio}
\alias{substitute_exofficio}
\title{substitute ex officio helper function}
\usage{
substitute_exofficio(title.text)
}
\description{
condenses ex officio abbreviations to a standardized form
}
