% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_finance}
\alias{substitute_finance}
\title{substitute finance helper function}
\usage{
substitute_finance(TitleTxt)
}
\description{
condenses finance abbreviations to a standardized form
financial is included
}
