% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_general}
\alias{substitute_general}
\title{substitute general helper function}
\usage{
substitute_general(TitleTxt)
}
\description{
condenses general abbreviations to a standardized form
}
