% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_governor}
\alias{substitute_governor}
\title{substitute governor helper function}
\usage{
substitute_governor(TitleTxt)
}
\description{
condenses governor abbreviations to a standardized form
also works on governance and government
}
