% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_information}
\alias{substitute_information}
\title{substitute information helper function}
\usage{
substitute_information(TitleTxt)
}
\description{
condenses info abbreviations to a standardized form
}
