% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_institute}
\alias{substitute_institute}
\title{substitute institute root helper function}
\usage{
substitute_institute(TitleTxt)
}
\description{
condenses institute/institutional abbreviations to a standardized form

also includes instructor standardization
}
