% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_intelligence}
\alias{substitute_intelligence}
\title{substitute intelligence helper function}
\usage{
substitute_intelligence(TitleTxt)
}
\description{
condenses intel abbreviations to a standardized form
}
