% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_manage}
\alias{substitute_manage}
\title{substitute manage root helper function}
\usage{
substitute_manage(TitleTxt)
}
\description{
condenses management/manager/managing
abbreviations to a standardized form
}
