% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_marketing}
\alias{substitute_marketing}
\title{substitute marketing helper function}
\usage{
substitute_marketing(TitleTxt)
}
\description{
condenses marketing abbreviations to a standardized form
}
