% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_member}
\alias{substitute_member}
\title{substitute member helper function}
\usage{
substitute_member(title.text)
}
\description{
condenses member abbreviations to a standardized form
}
