% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_miscellaneous}
\alias{substitute_miscellaneous}
\title{substitute miscellaneous helper function}
\usage{
substitute_miscellaneous(title.text)
}
\description{
substituting all the one's we missed/haven't generalized
}
