% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_officer}
\alias{substitute_officer}
\title{substitute officer helper function}
\usage{
substitute_officer(TitleTxt)
}
\description{
condenses officer abbreviations to a standardized form
}
