% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_philanthropy}
\alias{substitute_philanthropy}
\title{substitute philanthropy helper function}
\usage{
substitute_philanthropy(TitleTxt)
}
\description{
condenses philanthropy abbreviations to a standardized form
}
