% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_planning}
\alias{substitute_planning}
\title{substitute planning helper function}
\usage{
substitute_planning(TitleTxt)
}
\description{
condenses planning abbreviations to a standardized form
includes planned
}
