% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_president}
\alias{substitute_president}
\title{substitute president helper function}
\usage{
substitute_president(TitleTxt)
}
\description{
condenses president abbreviations to a standardized form
ignores "presiding"
}
