% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_programs}
\alias{substitute_programs}
\title{substitute programs helper function}
\usage{
substitute_programs(title.text)
}
\description{
condenses programs abbreviations to a standardized form
skips over programming
}
