% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_public}
\alias{substitute_public}
\title{substitute public helper function}
\usage{
substitute_public(title.text)
}
\description{
condenses public abbreviations to a standardized form
}
