% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_relations}
\alias{substitute_relations}
\title{substitute relations helper function}
\usage{
substitute_relations(TitleTxt)
}
\description{
condenses relations abbreviations to a standardized form
}
