% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_representative}
\alias{substitute_representative}
\title{substitute representative helper function}
\usage{
substitute_representative(title.text)
}
\description{
condenses rep abbreviations to a standardized form
}
