% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_secretary}
\alias{substitute_secretary}
\title{substitute secretary helper function}
\usage{
substitute_secretary(title.text)
}
\description{
condenses secretary abbreviations to a standardized form
}
