% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_senior}
\alias{substitute_senior}
\title{substitute seniority helper function}
\usage{
substitute_senior(TitleTxt)
}
\description{
condenses senior/junior abbreviations to a standardized form
}
