% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_strategy}
\alias{substitute_strategy}
\title{substitute strategy helper function}
\usage{
substitute_strategy(TitleTxt)
}
\description{
condenses strategy abbreviations to a standardized form
also condenses strategic to strategy
}
