% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_transportation}
\alias{substitute_transportation}
\title{substitute transportation helper function}
\usage{
substitute_transportation(TitleTxt)
}
\description{
condenses transportation abbreviations to a standardized form
}
