% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_treasurer}
\alias{substitute_treasurer}
\title{substitute treasurer helper function}
\usage{
substitute_treasurer(TitleTxt)
}
\description{
condenses treasurer abbreviations to a standardized form
}
