% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_vice}
\alias{substitute_vice}
\title{substitute vice helper function}
\usage{
substitute_vice(TitleTxt)
}
\description{
condenses vice president abbreviations to a standardized form
}
