% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{test_standardize_of}
\alias{test_standardize_of}
\title{test 'standardize_of' function}
\usage{
test_standardize_of(n = 250, x = NULL)
}
\arguments{
\item{n}{How many test cases to process.}

\item{x}{A vector of test titles, if NULL uses internal data.}
}
\value{
A data frame with the input titles (RAW) and processed titles (FIXED).
}
\description{
Generate a test table to show effects of the function.
}
