% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tinypartvii}
\alias{tinypartvii}
\title{Sample of 10,000 orgs (~81,000 titles) as a Part VII data frame.}
\description{
A data frame of raw IRS 990 Part VII compensation records drawn
from the 2023 NCCS efile release, in the current (modern) schema
(\code{OBJECTID}, \code{EIN2}, \code{RETURN_TYPE}, \code{ORG_NAME_L1/L2}, \verb{F9_07_COMP_DTK_*}).
Ready to pipe into \code{standardize_df()}. Regenerated by
\code{data-raw/make-tinypartvii.R}.
}
