% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{to_helper}
\alias{to_helper}
\title{standardize to helper function}
\usage{
to_helper(x)
}
\arguments{
\item{x}{A character vector of titles.}
}
\description{
does what standardize to but on a singular level
checks for to in between parentheticals and replaces with until
}
