% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunction-cleaning.R
\name{unmingle}
\alias{unmingle}
\title{unmingle function}
\usage{
unmingle(TitleTxt)
}
\description{
unmingles titles that are stuck together
e.g. separates "executivedirector" into "executive director"
}
