% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-categorize-titles.R
\name{add_features}
\alias{add_features}
\title{engineer new features from existing fields}
\usage{
add_features(df)
}
\description{
adds helpful context fields in comp dataset
like pay rank, hour rank, counts of types of employees, etc.
}
