% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-gen-status-codes.R
\name{add_status_flag}
\alias{add_status_flag}
\title{add a boolean status flag to the compensation dataframe}
\usage{
add_status_flag(df, title, s.code)
}
\arguments{
\item{df}{A compensation dataframe}

\item{title}{Which version of the title string to use (defaults to "TitleTxt6")}

\item{s.code}{Any of the unique status.qualifier strings from df.status (e.g. "FORMER")}
}
\description{
Search for variants of a status code and add a boolean flag
to the data frame, 1 if any status variant is present in the tite,
0 otherwise.
}
