% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{amp_helper}
\alias{amp_helper}
\title{ampersand standardization helper function}
\usage{
amp_helper(x)
}
\description{
operates on a singular title level
splits it, checks that if both/all sub titles are valid titles,
then returns a boolean whether and was used as a title separator
later lapply-ed for standardize_and usage

as opposed to and, its default is true
}
