% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{apply_substitutes}
\alias{apply_substitutes}
\title{apply fix spelling function}
\usage{
apply_substitutes(TitleTxt)
}
\description{
subwrapper function for standardize spelling by applying all the substitutes
on the titletxt vector (or singular string)
}
