% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{apply_substitutes}
\alias{apply_substitutes}
\title{apply substitutes wrapper function}
\usage{
apply_substitutes(title.text)
}
\description{
apply custom dictionary of common abbreviations and misspellings
to standardize titles
}
