% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-standardize-titles.R
\name{basic_csuite_fixes}
\alias{basic_csuite_fixes}
\title{basic c-suite fixes wrapper function}
\usage{
basic_csuite_fixes(
  comp.data,
  title = "TitleTxt6",
  hours = "TOT.HOURS",
  pay = "TOT.COMP",
  officer = "F9_07_COMP_DTK_POS_OFF_X"
)
}
\description{
applies conditional logic and creates some flags with helpful info
}
