% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-codes.R
\name{categorize_miscellaneous}
\alias{categorize_miscellaneous}
\title{categorize miscellaneous wrapper function}
\usage{
categorize_miscellaneous(comp.data)
}
\description{
categorizes, removes, and creates flags for:
schedule o, as needed, at large, ex officio, and co
also includes numericals
}
