% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-codes.R
\name{categorize_qualifiers}
\alias{categorize_qualifiers}
\title{categorize qualifiers function}
\usage{
categorize_qualifiers(comp.data)
}
\description{
takes in a data frame and categorizes the role statuses
future, former, interim, current (current is default, no status for that)
}
