% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-categorize-titles.R
\name{categorize_titles}
\alias{categorize_titles}
\title{categorize titles function}
\usage{
categorize_titles(comp.data)
}
\description{
categorizes the titles in comp.data
works on TitleTxt7
wrapper function
}
