% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-R.R
\name{categorize_titles}
\alias{categorize_titles}
\title{categorize titles function
DEPRECATED but to be updated}
\usage{
categorize_titles(comp.data)
}
\description{
categorizes the titles in comp.data
(based on TitleTxt2, so if named otherwise will have to rename)
note: copied from old-R/OLD-FUNCTIONS

to work well, it's helpful to remove the of's in titles
}
