% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-conditional-logic.R
\name{clean_up_ceos}
\alias{clean_up_ceos}
\title{clean up ceos function}
\usage{
clean_up_ceos(comp.data)
}
\arguments{
\item{comp.data}{A Part VII compensation data frame.}
}
\description{
removes all duplicate instances of ceos within a single org
}
