% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{condense_abbreviations}
\alias{condense_abbreviations}
\title{condense c-suite abbreviations function}
\usage{
condense_abbreviations(title.text)
}
\description{
condenses spaced out c-suite abbreviations
e.g. C E O --> CEO

also includes miscellaneous executive director and key employee substitutions
}
