% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-conditional-logic.R
\name{conditional_logic}
\alias{conditional_logic}
\title{conditional logic function}
\usage{
conditional_logic(comp.data)
}
\arguments{
\item{comp.data}{A Part VII compensation data frame.}
}
\description{
TODO formalize the conditional logic rules
(have a couple of them already thought out in categorize titles, but need
to migrate them)
}
