% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-standardize-titles.R
\name{conditional_logic}
\alias{conditional_logic}
\title{conditional logic wrapper function}
\usage{
conditional_logic(
  comp.data,
  title = "TitleTxt6",
  hours = "TOT.HOURS",
  pay = "TOT.COMP",
  officer = "F9_07_COMP_DTK_POS_OFF_X"
)
}
\description{
applies conditional logic and creates flags with helpful info
}
