% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-conditional-logic.R
\name{director_correction}
\alias{director_correction}
\title{director correction function}
\usage{
director_correction(df)
}
\arguments{
\item{df}{A Part VII compensation data frame.}
}
\description{
confirm all directors are properly mapped
correcting board members "directors" that are not actually directors
}
