% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_chair}
\alias{fix_chair}
\title{standardize versions of 'chair'}
\usage{
fix_chair(TitleTxt)
}
\description{
condenses chair abbreviations to a standardized form
e.g. chairperson,chairman,chairwoman --> chair
includes vice chair standardizations
}
