% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_comm}
\alias{fix_comm}
\title{standardize abbreviations for 'comm'}
\usage{
fix_comm(TitleTxt)
}
\description{
condenses communication and committee abbreviations to a standardized form
}
