% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_compliance}
\alias{fix_compliance}
\title{standardize versions of 'compliance'}
\usage{
fix_compliance(TitleTxt)
}
\description{
condenses compliance abbreviations to a standardized form
}
