% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_coordinator}
\alias{fix_coordinator}
\title{standardize versions of 'coordinator'}
\usage{
fix_coordinator(TitleTxt)
}
\description{
condenses coordinator abbreviations to a standardized form
note: instead of standardizing "CO" at the end of a title to coordinator,
it is instead mapped to "COMMITTEE"
}
