% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_corresponding}
\alias{fix_corresponding}
\title{standardize versions of 'corresponding'}
\usage{
fix_corresponding(TitleTxt)
}
\description{
condenses corresponding abbreviations to a standardized form
}
