% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_development}
\alias{fix_development}
\title{standardize versions of 'development'}
\usage{
fix_development(TitleTxt)
}
\description{
condenses development abbreviations to a standardized form
}
