% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{fix_double_and}
\alias{fix_double_and}
\title{double 'and' split rule}
\usage{
fix_double_and(x)
}
\description{
Identifies cases with the pattern "title AND word AND word"
and makes the substitution "title & word AND word".
}
