% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_enrollment}
\alias{fix_enrollment}
\title{standardize versions of 'enrollment'}
\usage{
fix_enrollment(TitleTxt)
}
\description{
condenses enrollment abbreviations to a standardized form
}
