% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_exofficio}
\alias{fix_exofficio}
\title{standardize versions of 'ex officio'}
\usage{
fix_exofficio(TitleTxt)
}
\description{
condenses ex officio abbreviations to a standardized form
}
