% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_governor}
\alias{fix_governor}
\title{standardize versions of 'governor'}
\usage{
fix_governor(TitleTxt)
}
\description{
condenses governor abbreviations to a standardized form
also works on governance and government
}
