% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_institute}
\alias{fix_institute}
\title{standardize versions of 'institute' (root of institutional)}
\usage{
fix_institute(TitleTxt)
}
\description{
condenses institute/institutional abbreviations to a standardized form

also includes instructor standardization
}
