% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_manage}
\alias{fix_manage}
\title{standardize versions of 'manage' (root of management)}
\usage{
fix_manage(TitleTxt)
}
\description{
condenses management/manager/managing
abbreviations to a standardized form
}
