% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{fix_misc_splits}
\alias{fix_misc_splits}
\title{misc split rules}
\usage{
fix_misc_splits(x)
}
\description{
Miscelaneous cases that should be split into
two titles but are not addressed by general rules.
}
