% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_miscellaneous}
\alias{fix_miscellaneous}
\title{standardize miscellaneous list of misspellings and abbreviations}
\usage{
fix_miscellaneous(TitleTxt)
}
\description{
substituting all the one's we missed/haven't generalized
}
