% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_philanthropy}
\alias{fix_philanthropy}
\title{standardize versions of 'philanthropy'}
\usage{
fix_philanthropy(TitleTxt)
}
\description{
condenses philanthropy abbreviations to a standardized form
}
