% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_president}
\alias{fix_president}
\title{standardize versions of 'president'}
\usage{
fix_president(TitleTxt)
}
\description{
condenses president abbreviations to a standardized form
ignores "presiding"
}
