% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_representative}
\alias{fix_representative}
\title{standardize versions of 'representative'}
\usage{
fix_representative(TitleTxt)
}
\description{
condenses rep abbreviations to a standardized form
}
