% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_spelling}
\alias{fix_spelling}
\title{fixes spelling wrapper function}
\usage{
fix_spelling(TitleTxt)
}
\description{
apply custom dictionary of common abbreviations and misspellings
to fix spelling of titles
operates on the title level instead of data frame and is used also in step 3
as well
}
