% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{fix_transportation}
\alias{fix_transportation}
\title{standardize versions of 'transportation'}
\usage{
fix_transportation(TitleTxt)
}
\description{
condenses transportation abbreviations to a standardized form
}
