% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-codes.R
\name{gen_status_codes}
\alias{gen_status_codes}
\title{generate status codes wrapper function}
\usage{
gen_status_codes(comp.data)
}
\description{
phase 4: remove all quantifiers and qualifiers
we create flags for all of the qualifiers too
the qualifiers include: schedule o, at large, as needed, ex officio, co,
ordinal numbers, and role statuses (former, future, current, and interim)

roughly 2.5 minutes for 100,000 titles
}
