% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-gen-taxonomy.R
\name{gen_taxonomy}
\alias{gen_taxonomy}
\title{generate taxonomy function}
\usage{
gen_taxonomy(comp.data)
}
\description{
TODO not written yet, but we will generate the taxonomy using the dynamic google sheets
no longer manipulating the titles
}
