% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-gen-status-codes.R
\name{get_variants}
\alias{get_variants}
\title{retrieve all status variants for a unique status code}
\usage{
get_variants(s.code)
}
\arguments{
\item{s.code}{Any of the unique status.qualifier strings from df.status ("FUTURE","FORMER","INTERIM",etc)}
}
\description{
Search for variants of a status code (stored in google sheets),
and combine them all into a single regex search string separated by OR.
}
