% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-split-titles.R
\name{identify_split_num}
\alias{identify_split_num}
\title{identify split num function}
\usage{
identify_split_num(x)
}
\description{
identifies if a split is in a title using the & separator
and returns the number of titles actually present
}
