% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{of_title_helper}
\alias{of_title_helper}
\title{standardize titles \emph{something} 'of' \emph{something}}
\usage{
of_title_helper(x)
}
\description{
finds instances where of should be in the title and replaces as necessary
}
