% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-classifier-functions.R
\name{pre_clean}
\alias{pre_clean}
\title{pre cleaning function}
\usage{
pre_clean(title.text)
}
\description{
gets rid of meaningless punctuation (like periods),
converts titles to uppercase
}
