% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-remove-dates.R
\name{remove_dates}
\alias{remove_dates}
\title{remove/clean dates wrapper function, takes in a data frame}
\usage{
remove_dates(comp.data)
}
\description{
cleans the dates by first identifying date, setting a flag if it finds one,
then removing the date if it is present
}
