% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-gen-status-codes.R
\name{remove_status}
\alias{remove_status}
\title{remove a status qualifier from the title}
\usage{
remove_status(x, s.code)
}
\arguments{
\item{x}{Vector of titles}

\item{s.code}{Any of the unique status.qualifier strings from df.status (e.g. "FORMER")}
}
\description{
Search for variants of a status code and remove them.
If the status code is the entire title (e.g. 'ex officio')
it is replaced with the standardized version of the status code.
}
