% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{slash_helper}
\alias{slash_helper}
\title{standardize slash helper function}
\usage{
slash_helper(x)
}
\description{
returns 0,1,2 depending on whether a slash is used as a title separator,
word separator, or just extraneous
operates on the atomic vector level
}
