% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{spellcheck}
\alias{spellcheck}
\title{spell check function}
\usage{
spellcheck(title.text)
}
\description{
correct small spelling mistakes
right now only correcting president, treasurer, and trustee misspellings

pretty slow function right now...
}
