% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-split-titles.R
\name{split_titles}
\alias{split_titles}
\title{split titles function}
\usage{
split_titles(comp.data, title = "TitleTxt3")
}
\description{
wrapper function:
splitting titles and returning a new data frame
it will be applied on a data frame

simply splitting on "&"
assuming it's already cleaned well

runtime on 100,000 titles ~ 45 minutes
}
