% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunction-cleaning.R
\name{split_titles}
\alias{split_titles}
\title{split titles wrapper function}
\usage{
split_titles(title.text)
}
\description{
basically does the same as split compound titles, but is more standardized
only separator is &, removes punctuation too
}
