% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunction-cleaning.R
\name{split_titles}
\alias{split_titles}
\title{split titles function}
\usage{
split_titles(comp.data)
}
\description{
basically doing the smae thing as build standard titles but on a split title
wrapper function
it will be applied on a data frame however (sorta like build_standard_titles

simply splitting on "&"
assuming it's already cleaned well,
this function is to take the place of split_compound_titles
}
