% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunction-cleaning.R
\name{standardize_comma}
\alias{standardize_comma}
\title{standardize comma function}
\usage{
standardize_comma(TitleTxt)
}
\description{
distinguish if comma is for separator, extraneous, or of
we also deal with commas as "of"(e.g. VP, finance = VP of Finance)
}
