% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{standardize_conj}
\alias{standardize_conj}
\title{split conjunctions function}
\usage{
standardize_conj(comp.data, title = "TitleTxt2")
}
\description{
wrapper for cleaning conjunctions (all the standardizations come in here)
}
