% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunction-cleaning.R
\name{standardize_of}
\alias{standardize_of}
\title{standardize "of" usage function}
\usage{
standardize_of(TitleTxt)
}
\description{
detects whether “of” is a part of the title, or separator
for example, TITLE OF TITLE is fine, “AS OF DATE” is not

if "of" is part of a date (and namely as of), then we replace with "SINCE"
for also gets mapped to of?

depending on context, we can add in "of" between vice president or director
and its subject
}
