% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-gen-status-codes.R
\name{standardize_qualifiers}
\alias{standardize_qualifiers}
\title{standardize qualifiers function}
\usage{
standardize_qualifiers(TitleTxt)
}
\description{
combining all mappings to standardize future, former, current, and interim
to their respective categories
Note: current is just treated as a regular title
(currently, current is just thrown away with no flag, but that can be changed)
}
