% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{standardize_spelling}
\alias{standardize_spelling}
\title{standardize spelling wrapper function}
\usage{
standardize_spelling(comp.data, title = "TitleTxt4")
}
\description{
apply custom dictionary of common abbreviations and misspellings
to standardize titles
}
