% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-standardize-titles.R
\name{standardize_titles}
\alias{standardize_titles}
\title{standardize titles function}
\usage{
standardize_titles(
  comp.data,
  officer = "F9_07_COMP_DTK_POS_OFF_X",
  gs_title_xwalk = NULL
)
}
\arguments{
\item{comp.data}{A Part VII compensation data frame.}

\item{officer}{Name of the officer-flag column.}

\item{gs_title_xwalk}{Optional title-standardization crosswalk; if \code{NULL}, loaded from the bundled snapshot.}
}
\description{
Maps multiple variants of a title onto the standard form
that is defined in the Google sheet:
\href{https://docs.google.com/spreadsheets/d/1iYEY2HYDZTV0uvu35UuwdgAUQNKXSyab260pPPutP1M/edit?usp=sharing}{Title standardization crosswalk}
}
