% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-standardize-titles.R
\name{standardize_titles}
\alias{standardize_titles}
\title{standardize titles function}
\usage{
standardize_titles(
  comp.data,
  title = "TitleTxt6",
  hours = "TOT.HOURS",
  pay = "TOT.COMP",
  officer = "F9_07_COMP_DTK_POS_OFF_X"
)
}
\description{
TODO not written yet, but we will generate the standardizations using the dynamic google sheets
still title manipulation, but now we need context of other codes
}
