% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-classifier-functions.R
\name{standardize_titles}
\alias{standardize_titles}
\title{standardize titles function
DEPRECATED}
\usage{
standardize_titles(title.text)
}
\description{
\code{standardize_titles} provides a hard-coded mapping from raw titles to their
standardized form. For example, "PRES" and "PRESIDE" will both be mapped to
"PRESIDENT", and "CEO" and "EX DIR" will both be mapped to "CEO". This function
takes in a string and outputs a string. As a note, for future work, it's
potentially possible to use agrep to allow for spelling mistakes in long titles,
and/or also narrow down the title range depending on sector to improve efficiency.
}
