% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-standardize-conj.R
\name{standardize_to}
\alias{standardize_to}
\title{standardize "to" function}
\usage{
standardize_to(TitleTxt)
}
\description{
detects whether to is part of a date or not
if it is, replace with "UNTIL", if not, leave be

Note: we have already removed dates by this point so we
cannot look for numbers as a means of detecting date

alternate way of standardize to is to check the date flag:
if we detected a date, then likely the "to" should be an until
}
