% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_academics}
\alias{substitute_academics}
\title{substitute academics helper function}
\usage{
substitute_academics(TitleTxt)
}
\description{
condenses academics abbreviations to a standardized form
}
