% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_admissions}
\alias{substitute_admissions}
\title{substitute admissions helper function}
\usage{
substitute_admissions(TitleTxt)
}
\description{
condenses admissions abbreviations to a standardized form
}
