% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_chair}
\alias{substitute_chair}
\title{substitute chair helper function}
\usage{
substitute_chair(TitleTxt)
}
\description{
condenses chair abbreviations to a standardized form
e.g. chairperson,chairman,chairwoman --> chair
includes vice chair standardizations
}
