% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_compliance}
\alias{substitute_compliance}
\title{substitute compliance helper function}
\usage{
substitute_compliance(title.text)
}
\description{
condenses compliance abbreviations to a standardized form
}
