% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_intelligence}
\alias{substitute_intelligence}
\title{substitute intelligence helper function}
\usage{
substitute_intelligence(title.text)
}
\description{
condenses intel abbreviations to a standardized form
}
