% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_operations}
\alias{substitute_operations}
\title{substitute operations helper function}
\usage{
substitute_operations(title.text)
}
\description{
condenses operations abbreviations to a standardized form
function leaves "operating" alone --> does not convert to operations
}
