% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_planning}
\alias{substitute_planning}
\title{substitute planning helper function}
\usage{
substitute_planning(title.text)
}
\description{
condenses planning abbreviations to a standardized form
includes planned
}
