% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_programs}
\alias{substitute_programs}
\title{substitute programs helper function}
\usage{
substitute_programs(TitleTxt)
}
\description{
condenses programs abbreviations to a standardized form
skips over programming
}
