% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_projects}
\alias{substitute_projects}
\title{substitute projects helper function}
\usage{
substitute_projects(TitleTxt)
}
\description{
condenses projects abbreviations to a standardized form
}
