% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_representative}
\alias{substitute_representative}
\title{substitute representative helper function}
\usage{
substitute_representative(TitleTxt)
}
\description{
condenses rep abbreviations to a standardized form
}
