% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-substitutions.R
\name{substitute_systems}
\alias{substitute_systems}
\title{substitute systems helper function}
\usage{
substitute_systems(title.text)
}
\description{
condenses systems abbreviations to a standardized form
}
