% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-standardize-spelling.R
\name{substitute_technology}
\alias{substitute_technology}
\title{substitute technology helper function}
\usage{
substitute_technology(TitleTxt)
}
\description{
condenses tech abbreviations to a standardized form
}
